refactor: implemented new configuration for socket connections/reconnections
