+++ client/store/user.js
@@ -1,7 +1,38 @@
 import axios from 'axios'
 import history from '../history'
+import store from '.'
+import openTeacherSocket from '../socket/teacher'
+import openStudentSocket from '../socket/student'
 
 let socket
+
+//exported functions for sending reset actions to the store
+
+//if confirmed, sends start message to server with teacher's ID and details for the session
+export const startSession = (teacherId, sessionDetails) => {
+  if (window.confirm('Are you ready to start the session?')) {
+    socket.emit('start-session', teacherId, sessionDetails)
+
+    document.getElementById('start').hidden = true
+    document.getElementById('end').hidden = false
+
+    store.dispatch(resetSessionData())
+    store.dispatch(resetStudentData())
+  }
+}
+
+//if confirmed, sends message to server to end the session
+export const endSession = () => {
+  if (window.confirm('Are you sure you want to end the session?')) {
+    socket.emit('end-session')
+
+    document.getElementById('start').hidden = false
+    document.getElementById('end').hidden = true
+    document.getElementsByClassName('re-invite').forEach(button => {
+      button.parentNode.removeChild(button)
+    })
+  }
+}
 
 /**
  * ACTION TYPES
@@ -30,10 +61,11 @@ export const me = () => async dispatch => {
 
     //on loading up the session user, import the appropriate socket functionality, and send reconnect message to server
     if (res.data) {
-      if (res.data.isTeacher) socket = require('../socket/teacher').default
-      else socket = require('../socket/student').default
-
-      socket.emit('reconnect', res.data.id)
+      do {
+        let user = store.getState().user
+        if (user.id)
+          socket = user.isTeacher ? openTeacherSocket() : openStudentSocket()
+      } while (!socket)
 
       history.push('/session')
     }
@@ -54,10 +86,11 @@ export const auth = (email, password, method) => async dispatch => {
     dispatch(getUser(res.data))
 
     //on login/signup, import the appropriate socket functionality and send reconnect message to server
-    if (res.data.isTeacher) socket = require('../socket/teacher').default
-    else socket = require('../socket/student').default
-
-    socket.emit('reconnect', res.data.id)
+    do {
+      let user = store.getState().user
+      if (user.id)
+        socket = user.isTeacher ? openTeacherSocket() : openStudentSocket()
+    } while (!socket)
 
     history.push('/session')
   } catch (dispatchOrHistoryErr) {
@@ -71,7 +104,10 @@ export const logout = () => async dispatch => {
     dispatch(removeUser())
 
     //On logout, send logout message to the server
-    if (socket) socket.emit('logout', user)
+    do {
+      let user = store.getState().user
+      if (!user.id && socket) socket.emit('logout', user)
+    } while (user.id)
 
     history.push('/')
   } catch (err) {
